feat(proposal): setup sort algorithm for ES queries

Closes #298

See merge request makeorg/platform/core-api!468

+++ api/src/main/scala/org/make/api/tagtype/TagTypeResponse.scala
@@ -31,7 +31,7 @@ final case class TagTypeResponse(
   @(ApiModelProperty @field)(dataType = "string", example = "12345678-90ab-cdef-1234-567890abcdef") id: TagTypeId,
   label: String,
   @(ApiModelProperty @field)(dataType = "string", example = "DISPLAYED or HIDDEN") display: TagTypeDisplay,
-  weight: Float
+  weight: Int
 )
 
 object TagTypeResponse {
